Refactored Builder class thanks to the addition of the GameTool protocol
The Builder class used to wrap all of the game's functionality from the point of view of interaction, containing the logic for placement, destruction, and manipulation of Game Pieces. Since it had grown to a point it was hard to modify and extend, it has been refactored.
The builder class still exists but its role is to handle the logic of tools. Tools follow the GameTool protocol. Right now three tools exist (GameToolBuilder, GameToolDestroyer, GameToolManipulator). All these have a number of methods\events (as defined by the protocol) which are called accordingly by Builder when necessary.  In order to change functionality, it is merely necessary to switch out a tool for another, and new tools can be easily added by creating new protocol-conforming classes.