First implementation of aimed ARNPlaneDetection.
Rather than placing the plane where an Anchor is found, now GameToolARDetector places it at the exact point where the player is aiming. This allows for the player two switch between surfaces before deciding which one to use, as well as deciding exactly where the grid should go. It also allows for the implementation of further features, like rotation of the plane.