Overhauled GameTool and added plane rotation
GameTool has seen the introduction of onReselect(), which replaces the previous behaviour of calling onTap() when a tool that was already selected is selected again. onTap has been changed to onTap(_ sender: UITapGestureRecognizer) as well, and onPan(_ sender: UIPanGestureRecognizer) has been introduced. GameTool is now an objective-c protocol and all its methods are optional, which is good since no current GameTool makes use of all of them.

During AR plane detection, the plane can be rotated by swiping left and right. This means the user now has total control of both the position, and the orientation, of the plane!